Do not include offset time origins when a single time origin is requested
This affects exponential block sampling (block_period!=1). If the
offset time origins are included, there are more possible time
differences to compute a correlation function, but the results are
often "jumpy".

If a single time origin is requested, the expected behavior is to
average starting from the initial configuration only. Therefore offset
time origins must not be included.