g(r): crop particles on the border of the bounding cell when it is not periodic
The purpose is to compute g(r) in a subsystem of a bigger sample. This is now achieved by bounding the subsystem with a non-periodic simulation cell (requires the "periodic" cell attribute) and then cropping the linked cells that are on the border of the bounding cell. This only works with linked cells at the moment, thus passing a non-zero rmax is required.